add delete

+++ src/client/src/views/ParkingLot/ParkingLot.vue
@@ -27,6 +27,8 @@
             <div v-else class="Busy">
                 {{site.name.charAt(0).toUpperCase()}}{{ site.name.slice(1).toLowerCase()}}
             </div>
+            <p class="dateText">{{site.date.hour}}</p>
+            <p class="dateText">{{site.date.fecha}}</p>
         </li>
     </ul>
 </template>
@@ -44,14 +46,17 @@ export default {
         const listPark = useParking();
         listPark.getData()
         const deletePark = async (idPark)=>{
-            try{
-                const url = 'http://localhost:5000/sites/park/'+idPark;
-                const del = await axios.delete(url)
-                console.log(del);
-            }catch{
-                err => console.log(err);
-            }
-        }
+            const confirmar = confirm('Seguro?');
+            if (confirmar){
+                try{
+                    const url = 'http://localhost:5000/sites/park/'+idPark;
+                    const del = await axios.delete(url)
+                    console.log(del);
+                    listPark.getData()
+                }catch{
+                    err => console.log(err);
+                }
+        }}
         return{
             listPark,deletePark
         }
@@ -60,10 +65,14 @@ export default {
 </script>
 
 <style scoped>
+.dateText{
+    margin: 0.5vw auto;
+    font-size: 1vw;
+}
 .crossDelete{
     width: 1vw;
-    height: 1vw;
-    font-size: 0.7vw;
+    height: 2vw;
+    font-size: 1vw;
     background-color: rgb(190, 7, 7);
     margin: 0vw 1vw;
     padding: 0;
@@ -104,7 +113,7 @@ export default {
 }
 .cartSite{
     text-align: center;
-    padding: 1.5vw 0vw;
+    padding: 0.5vw 0vw;
     width: 10vw;
     background-color: rgb(87, 87, 87);
     border-radius: 0.5vw;
